Create Strapi bootleg
Strapi is a now a relic of the past.
