+++ server/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "server",
+  "version": "0.0.1",
+  "description": "Strapi Bootleg",
+  "main": "index.ts",
+  "scripts": {
+    "start": "npx ts-node ./src/index.ts",
+    "dev": "npx nodemon"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@types/express": "^4.17.21",
+    "cors": "^2.8.5",
+    "express": "^4.18.2",
+    "ts-node": "^10.9.2"
+  }
+}
