+++ src/pages/events.jsx
@@ -1,15 +1,15 @@
 import EventAttractionTemplate from "@/components/EventAttractionTemplate";
 import React from "react";
 import { useEffect, useState } from 'react'
-
+import { SERVER_URL } from "./index";
 export default function events() {
 	const [strapiData, setStrapiData] = useState();
 
 	useEffect(() => {
 		async function fetchStrapiData() {
-			const response = await fetch(`https://strapi.discoverlincoln-t2-c8.civiconnect.net/api/events?populate=*`)
+			const response = await fetch(`${SERVER_URL}/api/events`)
 			const data = await response.json()
-			setStrapiData(data.data)
+			setStrapiData(data)
 		}
 
 		fetchStrapiData()
