--- src/pages/homepage.jsx
@@ -1,755 +1,0 @@
-import React from "react";
-import Image from "next/image";
-
-import TextInput from "@/components/TextInput";
-import Section from "@/components/Section";
-import LargeCardMobile from "@/components/LargeCardMobile";
-import CardCarousel from "@/components/CardCarousel";
-import Gallery from "@/components/Gallery";
-import GalleryImage from "@/components/GalleryImage";
-import SmallCard from "@/components/SmallCard";
-import DefaultButton from "@/components/DefaultButton";
-import Footer from "@/components/Footer";
-import Navbar from "@/components/Navbar";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faMagnifyingGlass } from "@fortawesome/free-solid-svg-icons";
-import { useEffect, useState } from "react";
-
-export default function Homepage() {
-  const [eventsStrapiData, setEventsStrapiData] = useState([]); // events
-  const [attractionStrapiData, setAttractionStrapiData] = useState([]);
-  const [businessStrapiData, setBusinessStrapiData] = useState([]);
-  const [galleryStrapiData, setGalleryStrapiData] = useState([]);
-
-  const [searchQuery, setSearchQuery] = useState([]); // the thing to search for
-
-  useEffect(() => {
-    async function fetchStrapiData() {
-      const eventResponse = await fetch(
-        "https://strapi.discoverlincoln-t2-c8.civiconnect.net/api/events?populate=*"
-      );
-      const eventData = await eventResponse.json();
-      setEventsStrapiData(eventData.data);
-
-      const attractionResponse = await fetch(
-        "https://strapi.discoverlincoln-t2-c8.civiconnect.net/api/attrractions?populate=*"
-      );
-      const attractionData = await attractionResponse.json();
-      setAttractionStrapiData(attractionData.data);
-
-      const businessResponse = await fetch(
-        "https://strapi.discoverlincoln-t2-c8.civiconnect.net/api/businesses?populate=*"
-      );
-      const businessData = await businessResponse.json();
-      setBusinessStrapiData(businessData.data);
-
-      const galleryResponse = await fetch(
-        "https://strapi.discoverlincoln-t2-c8.civiconnect.net/api/gallery?populate=*"
-      );
-      const galleryData = await galleryResponse.json();
-      setGalleryStrapiData(galleryData.data.attributes.images.data);
-    }
-
-    fetchStrapiData();
-  }, []);
-
-  /**
-   * Read from text input component
-   * @param {*} data
-   */
-  const readQuery = (data) => {
-    setSearchQuery(data)
-  }
-
-  /**
-   * Take action on enter press
-   * @param {*} event 
-   */
-  const handleKeyPress = (keyPress) => {
-    if (keyPress.key === 'Enter') {
-      // I am not sure how well this works with % encoding but oh well
-      window.location.href = `/search?searchQuery=${searchQuery}`
-    }
-  }
-
-  return (
-    <>
-      <div className="homepage">
-        <div className="everything-excl-foot">
-          <Navbar isHomepage={true}></Navbar>
-          <div className="landing-view-bg">
-            <div className="landing-view-wrap">
-              <div className="landing-view-greet">
-                <div className="greeting-frame">
-                  <h1 className="discover-text">Discover</h1>
-                  <h1 className="lincoln-text">LINCOLN</h1>
-                  <p className="greeting-description-text">
-                    Welcome to the Town of Lincoln! The places you will go, and
-                    the sights you will see, may the rolling hills be in your
-                    favour.
-                  </p>
-                  <div className="search-area">
-                    <div className="see-lincoln-wrapper">
-                      <DefaultButton
-                        className="see-lincoln"
-                        fontSize={"18px"}
-                        bgColor={"white"}
-                        textColor={"black"}
-                        padding={"0px 12px"}
-                      >
-                        SEE LINCOLN
-                      </DefaultButton>
-                    </div>
-                    <div className="search-button-area">
-                      <DefaultButton isLink={true} href={{ pathname: "./search", query: { searchQuery } }} className="see-lincoln home-search-btn">
-                        <FontAwesomeIcon icon={faMagnifyingGlass} />
-                      </DefaultButton>
-                      <TextInput
-                        type="text"
-                        className="search-input"
-                        placeholder="Search..."
-                        padding={"0px 12px"}
-                        dataOut={readQuery}
-                        onKeyDownOut={handleKeyPress}
-                      ></TextInput>
-                    </div>
-                  </div>
-                </div>
-                <div className="upcoming-events-wrap">
-                  <div className="event-card-frame">
-                    {/* <h2>INSERT SMALL EVENT CARDS</h2> */}
-                    <div className="upcoming-events-gap-desktop"></div>
-                    <div className="upcoming-events-gap-mobile"></div>
-                    <CardCarousel margin="0px 0px 20px 0px">
-                      <h1 className="upcoming-events-subtitle">
-                        Upcoming Events
-                      </h1>
-
-                      {eventsStrapiData?.map((card, index) => (
-                        <li key={index} className="smallCardli">
-                          <SmallCard
-                            title={card.attributes.title}
-                            imgSrc={card.attributes.image.data.attributes.url}
-                            category={card.attributes.tags}
-                          >
-                            {" "}
-                            <Image
-                              src="Icon-glass.svg"
-                              width={20}
-                              height={20}
-                              alt="uwu"
-                            />
-                          </SmallCard>
-                        </li>
-                      ))}
-                    </CardCarousel>
-                  </div>
-                  <DefaultButton
-                    isLink={true}
-                    href="./events"
-                    className="homepage-SmallCard"
-                    padding={"0px 12px"}
-                  >
-                    SEE ALL
-                  </DefaultButton>
-                </div>
-              </div>
-              <div className="topology-wrap">
-                <svg
-                  className="topology"
-                  viewBox="0 0 1440 226"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <g clip-path="url(#clip0_18_5179)">
-                    <path
-                      d="M0 119.831V10.6616H1.774C2.75276 10.6616 7.21835 10.9755 11.7451 11.3941C21.7774 12.3009 31.0144 12.6497 55.3611 13.1031C74.3246 13.4868 93.1045 13.8355 126.321 14.4982C184.496 15.6492 179.48 15.3702 187.799 17.5326C196.119 19.6602 202.542 23.1481 211.106 30.1237C212.574 31.3445 214.104 32.6001 214.41 32.914C217.285 35.6345 219.67 38.076 221.444 40.1338C224.503 43.6216 227.072 44.947 232.578 45.8887C235.025 46.3422 237.594 46.7956 238.267 46.9351C239.307 47.1792 243.65 47.9465 254.172 49.7602C255.212 49.9346 259.432 50.7019 263.653 51.4693C267.874 52.2366 271.606 52.8644 272.034 52.8644C272.523 52.8644 273.319 53.0039 273.93 53.1783C274.481 53.3527 277.845 54.0154 281.393 54.6083C284.941 55.2361 288.673 55.8639 289.652 56.0383C290.692 56.2127 294.913 56.98 299.133 57.7474C303.354 58.5147 307.086 59.1425 307.514 59.1425C308.003 59.1425 308.799 59.282 309.41 59.4564C310.267 59.7354 316.445 60.8515 325.132 62.3164C326.172 62.4908 329.475 63.0838 332.472 63.6069C335.531 64.165 339.201 64.8277 340.731 65.0718C342.26 65.316 344.034 65.6299 344.707 65.7694C345.38 65.9089 348.683 66.5367 352.048 67.1296C355.412 67.7575 360.122 68.7689 362.447 69.3967C364.833 70.0246 367.83 70.8267 369.176 71.1755C370.522 71.5243 374.681 72.6055 378.352 73.617C382.083 74.6285 386.182 75.7097 387.528 76.0585C396.398 78.2558 409.917 82.1273 436.099 89.8703C442.338 91.7189 445.03 93.1489 447.721 96.1135C448.822 97.3692 450.902 99.6711 452.309 101.206C456.53 105.775 459.772 107.623 463.871 107.623C464.788 107.623 467.174 107.24 469.193 106.751C471.946 106.089 473.903 105.879 477.329 105.879C482.284 105.879 483.446 106.124 488.768 108.321C490.603 109.053 494.335 110.553 497.026 111.634C499.718 112.716 504 114.425 506.508 115.471C509.016 116.517 512.992 118.122 515.317 119.029C517.641 119.935 520.394 121.052 521.434 121.47C526.144 123.423 527.735 124.086 529.937 124.888C533.424 126.248 535.626 128.271 538.012 132.387C539.174 134.48 540.153 136.433 540.153 136.712C540.153 137.026 540.398 137.549 540.765 137.898C541.56 138.735 545.047 145.152 545.047 145.815C545.047 146.094 545.291 146.617 545.658 146.966C546.454 147.803 549.941 154.221 549.941 154.884C549.941 155.163 550.185 155.686 550.491 156.035C550.797 156.348 551.653 157.395 552.387 158.302C553.794 160.046 556.119 161.092 560.34 161.999C561.502 162.243 563.888 162.836 565.539 163.254C568.476 164.022 570.8 164.58 575.205 165.661C576.245 165.905 578.814 166.393 580.833 166.742C582.851 167.056 585.054 167.475 585.726 167.614C586.889 167.858 601.02 170.404 605.607 171.172C606.77 171.381 608.299 171.66 608.972 171.8C610.195 172.079 624.877 174.694 628.853 175.357C630.015 175.566 631.545 175.845 632.218 175.985C633.441 176.264 648.122 178.88 652.099 179.542C653.261 179.752 654.79 180.031 655.463 180.17C656.687 180.449 671.368 183.065 675.344 183.728C676.506 183.937 678.036 184.216 678.709 184.356C679.81 184.6 693.268 187.006 698.59 187.913C699.752 188.123 701.281 188.402 701.954 188.541C703.178 188.82 717.859 191.436 721.835 192.099C722.997 192.308 724.527 192.587 725.2 192.727C726.423 193.006 741.41 195.691 744.775 196.249C745.815 196.424 748.262 196.842 750.28 197.226C761.842 199.388 768.938 199.249 776.707 196.703C779.521 195.796 796.16 189.378 808.7 184.356C822.77 178.74 841.733 171.416 845.098 170.3C849.563 168.8 853.662 168.207 867.12 167.056C888.836 165.208 897.4 165.173 908.105 166.777C919.055 168.416 919.483 168.416 924.377 167.579C932.819 166.114 940.588 167.23 942.79 170.265C943.952 171.869 945.115 177.554 945.115 181.844C945.115 184.635 946.399 189.797 947.378 190.738C948.54 191.994 952.088 193.459 969.89 200.156C980.656 204.236 990.076 207.759 990.811 208.003C991.85 208.317 997.784 208.422 1014.06 208.422C1032.16 208.422 1036.38 208.526 1038.4 208.98C1039.75 209.259 1041.28 209.573 1041.77 209.677C1042.99 209.922 1045.44 212.119 1045.44 212.956C1045.44 213.305 1045.74 213.723 1046.05 213.828C1046.36 213.933 1046.66 214.665 1046.66 215.397C1046.66 216.165 1046.91 217.455 1047.27 218.257C1047.58 219.095 1048.31 220.978 1048.86 222.478C1049.6 224.64 1050.21 225.408 1051.92 226.419C1053.08 227.117 1054 227.988 1054 228.337C1054 229 1040.36 229 527.001 229H0L0 119.831Z"
-                      fill="#005837"
-                    />
-                    <g filter="url(#filter0_d_18_5179)">
-                      <path
-                        d="M1344.26 302.64C1339.55 302.396 1332.15 301.978 1327.74 301.768C1312.33 301.036 1303.21 299.745 1292.87 296.885C1290.79 296.327 1288.84 295.839 1288.53 295.839C1288.22 295.839 1286.94 295.525 1285.66 295.141C1284.43 294.758 1281.5 294.025 1279.23 293.502C1276.97 292.979 1273.91 292.212 1272.44 291.758C1270.97 291.305 1269.51 290.956 1269.2 290.956C1268.89 290.956 1267.06 290.538 1265.16 290.049C1263.27 289.561 1260.7 288.898 1259.53 288.619C1258.37 288.34 1255.74 287.643 1253.72 287.119C1251.7 286.596 1249.38 286.038 1248.52 285.899C1247.67 285.759 1246.69 285.515 1246.32 285.341C1245.95 285.166 1244.61 284.818 1243.26 284.538C1241.98 284.259 1238.73 283.492 1236.1 282.795C1233.53 282.097 1231.27 281.539 1231.09 281.539C1230.9 281.539 1228.95 281.051 1226.68 280.458C1224.36 279.865 1221.97 279.272 1221.3 279.097C1220.63 278.923 1218.3 278.33 1216.1 277.737C1209 275.889 1202.82 274.877 1200.32 275.121C1197.56 275.4 1193.4 276.447 1190.1 277.702C1184.04 279.969 1179.39 280.667 1171.75 280.388C1161.9 280.074 1158.17 278.505 1148.32 270.657C1142.02 265.634 1136.69 262.74 1132.6 262.147C1131.68 262.007 1129.72 261.554 1128.19 261.1C1126.72 260.647 1125.19 260.263 1124.77 260.263C1124.34 260.263 1122.38 259.775 1120.42 259.217C1118.46 258.624 1116.51 258.17 1116.14 258.17C1115.77 258.17 1114.49 257.891 1113.33 257.508C1112.16 257.159 1110.51 256.74 1109.66 256.601C1108.8 256.461 1106.41 255.868 1104.27 255.345C1102.19 254.787 1100.3 254.334 1100.17 254.334C1100.05 254.334 1098.09 253.88 1095.89 253.287C1093.69 252.729 1091.73 252.241 1091.61 252.241C1091.49 252.241 1089.53 251.788 1087.33 251.195C1085.12 250.637 1083.17 250.148 1083.04 250.148C1082.92 250.148 1080.96 249.695 1078.76 249.102C1076.56 248.544 1074.42 248.056 1073.99 248.056C1073.56 248.056 1072.95 247.916 1072.64 247.742C1072.28 247.602 1067.63 246.172 1062.24 244.568C1056.86 242.998 1052.21 241.603 1051.84 241.429C1051.54 241.289 1047.81 240.173 1043.59 238.952C1039.43 237.767 1033.61 236.023 1030.74 235.151C1016.3 230.651 995.258 224.478 990.241 223.257C978.068 220.258 966.077 218.479 953.353 217.712C949.988 217.537 940.384 217.188 931.942 217.014C912.733 216.561 893.707 216.072 875.844 215.619C861.529 215.27 855.901 215.375 833.756 216.526C819.38 217.293 798.702 217.293 790.627 216.526C784.693 215.933 776.802 214.852 770.44 213.735C763.771 212.55 754.289 210.945 736.487 207.911C726.21 206.167 715.198 204.318 712.017 203.76C695.806 200.97 693.053 200.517 689.077 199.854C686.752 199.47 684.122 199.017 683.265 198.877C682.409 198.703 679.656 198.249 677.148 197.796C669.195 196.471 665.341 195.703 660.141 194.308C657.388 193.541 655.002 192.948 654.819 192.948C654.635 192.948 651.699 192.181 648.273 191.239C644.786 190.332 640.32 189.111 638.302 188.588C636.283 188.065 629.615 186.321 623.497 184.717C617.319 183.112 612.18 181.787 611.996 181.787C611.874 181.787 607.898 180.775 603.309 179.52C592.237 176.52 590.524 176.067 583.244 174.218C579.879 173.346 575.23 172.126 572.844 171.498C565.809 169.614 559.08 167.87 553.268 166.371C550.271 165.603 545.56 164.383 542.869 163.65C540.177 162.918 536.873 162.081 535.528 161.732C534.182 161.418 530.328 160.372 526.963 159.5C516.196 156.605 508.427 154.582 508.06 154.582C507.203 154.582 500.78 151.757 494.234 148.478C490.38 146.56 486.098 144.432 484.691 143.77C483.345 143.107 477.167 140.038 470.988 136.968C453.737 128.388 450.127 126.679 447.741 126.016C442.786 124.586 440.951 124.517 411.403 124.9C395.865 125.11 381.489 125.354 379.348 125.458C374.331 125.703 370.6 126.47 366.379 128.039C360.2 130.376 357.447 130.865 350.534 130.865C347.659 130.865 343.927 130.655 342.214 130.376C338.605 129.818 333.895 128.632 310.098 122.25C306.917 121.378 303.185 120.436 301.839 120.157C300.493 119.878 298.597 119.424 297.557 119.18C296.578 118.901 294.376 118.448 292.663 118.134C291.011 117.82 289.053 117.436 288.38 117.262C284.649 116.32 272.597 114.855 249.718 112.553C236.198 111.193 233.445 110.67 229.469 108.717C227.083 107.531 219.436 102.334 212.095 96.8582C203.531 90.4754 192.764 86.3947 179.367 84.4764C173.677 83.6393 160.402 83.2905 114.949 82.7324C39.704 81.8256 26.9185 81.6163 21.3515 81.1629C16.0293 80.7095 6.54719 79.6283 2.08141 78.9656L0.246163 78.6866L0.0626379 39.3787C0.00146286 17.7541 0.184988 0.0708008 0.429688 0.0708008C0.674388 0.0708008 2.44846 0.83812 4.28371 1.74496C6.05779 2.65179 8.44361 3.80278 9.48359 4.2562C10.5236 4.7445 15.4787 7.08135 20.4339 9.48795C35.9112 16.9868 36.89 17.3007 44.2922 17.6844C51.3885 18.0331 54.2025 18.7307 68.2728 23.4742C75.3079 25.811 84.0559 28.7408 87.7264 29.9964C91.4581 31.252 95.6792 32.6472 97.2086 33.1355C98.6768 33.6238 102.408 34.8794 105.406 35.9257C111.707 38.0882 117.213 39.588 121.067 40.1809C124.309 40.6692 133.424 40.704 137.768 40.2158C139.48 40.0065 142.6 39.6926 144.619 39.5182C146.638 39.3089 151.348 38.8206 155.019 38.437C158.751 38.0533 164.929 37.4255 168.783 37.0418C172.637 36.6582 177.348 36.2048 179.183 35.9955C181.018 35.8211 185.729 35.3328 189.583 34.9491C193.437 34.5655 197.964 34.0772 199.554 33.9028C205.794 33.1703 211.3 33.4842 222.617 35.0887C262.32 40.8785 264.339 41.1924 269.11 42.6921C272.597 43.8082 274.8 44.7499 286.545 49.9468C288.503 50.8188 290.95 51.9 292.051 52.3883C293.152 52.8766 295.599 53.9578 297.557 54.8298C299.514 55.7017 301.961 56.783 303.062 57.2713C304.164 57.7596 306.611 58.8408 308.568 59.7127C310.526 60.5847 312.973 61.6659 314.074 62.1542C315.175 62.6425 317.622 63.7238 319.58 64.5957C330.775 69.5833 333.65 70.3855 347.414 72.5829C349.923 73.0014 359.221 74.4663 368.03 75.8963C376.84 77.3263 386.444 78.6517 389.442 78.8958C393.234 79.1749 405.225 79.2446 430.001 79.1749L465.176 79.0354L490.809 76.5241C509.528 74.6755 517.909 74.0129 521.885 74.0129C527.881 74.0129 526.29 73.7687 555.348 79.1051C575.964 82.9417 577.065 83.0812 583.428 83.0812C588.995 83.0812 590.218 82.9417 599.15 81.3722C602.208 80.8141 606.735 80.0468 609.243 79.6283C611.752 79.1749 615.361 78.547 617.196 78.1983C627.107 76.3148 633.897 75.408 638.424 75.408C640.198 75.408 652.188 76.28 665.035 77.3263C692.747 79.6283 697.458 79.8375 703.881 79.2446C708.163 78.861 718.318 77.2565 721.5 76.4892C723.457 76.0009 727.984 75.0243 730.982 74.5012C732.511 74.2221 734.285 73.8385 734.958 73.699C739.852 72.5131 750.435 71.0831 761.263 70.1762C787.385 68.0138 787.385 68.0138 794.91 68.1882C800.66 68.3277 803.78 68.6416 809.286 69.5135C815.648 70.5599 817.116 70.6646 824.579 70.6646C829.106 70.6646 838.772 70.4901 845.991 70.246C853.209 70.0367 875.416 69.6531 895.237 69.4787C915.118 69.2694 941.546 68.9555 954.026 68.7811C976.293 68.4672 983.818 68.6067 991.587 69.5135C995.074 69.9321 1009.02 72.5131 1030.74 76.8031C1037.65 78.1634 1045.6 79.7329 1048.48 80.291C1067.14 83.8834 1068.55 84.2671 1082.43 89.6384C1088.12 91.8706 1099.87 96.3699 1108.43 99.6484C1131.19 108.368 1135.96 110.251 1138.84 111.542C1143.97 113.949 1148.26 118.029 1150.7 122.843C1152.36 126.191 1155.23 128.284 1160.06 129.748C1162.75 130.551 1163.92 130.69 1168.69 130.69C1173.77 130.69 1174.44 130.62 1177.87 129.539C1179.82 128.876 1182.21 128.109 1183.07 127.76C1185.14 126.958 1188.94 126.016 1192.79 125.319C1195.73 124.796 1203.19 124.831 1207.05 125.424C1208.82 125.703 1214.02 126.609 1218.55 127.481C1252.74 133.969 1265.96 135.329 1301.74 136.096C1319.24 136.48 1329.03 136.864 1342.43 137.666C1361.82 138.817 1395.34 137.875 1409.35 135.782C1410.57 135.573 1414.06 135.085 1417.06 134.701C1420.12 134.283 1423.24 133.829 1424.09 133.655C1424.95 133.515 1426.85 133.167 1428.38 132.922C1429.91 132.678 1431.44 132.399 1431.74 132.26C1432.11 132.12 1433.58 131.841 1435.11 131.562C1436.64 131.318 1438.1 130.969 1438.41 130.795C1438.78 130.655 1439.27 130.516 1439.51 130.516C1439.76 130.516 1440 169.196 1440 216.421V302.361L1426.11 302.571C1394.18 303.129 1353.07 303.164 1344.26 302.64Z"
-                        fill="#9ACA3C"
-                      />
-                    </g>
-                    <path
-                      d="M0 119.831V10.6616H1.774C2.75276 10.6616 7.21835 10.9755 11.7451 11.3941C21.7774 12.3009 31.0144 12.6497 55.3611 13.1031C74.3246 13.4868 93.1045 13.8355 126.321 14.4982C184.496 15.6492 179.48 15.3702 187.799 17.5326C196.119 19.6602 202.542 23.1481 211.106 30.1237C212.574 31.3445 214.104 32.6001 214.41 32.914C217.285 35.6345 219.67 38.076 221.444 40.1338C224.503 43.6216 227.072 44.947 232.578 45.8887C235.025 46.3422 237.594 46.7956 238.267 46.9351C239.307 47.1792 243.65 47.9465 254.172 49.7602C255.212 49.9346 259.432 50.7019 263.653 51.4693C267.874 52.2366 271.606 52.8644 272.034 52.8644C272.523 52.8644 273.319 53.0039 273.93 53.1783C274.481 53.3527 277.845 54.0154 281.393 54.6083C284.941 55.2361 288.673 55.8639 289.652 56.0383C290.692 56.2127 294.913 56.98 299.133 57.7474C303.354 58.5147 307.086 59.1425 307.514 59.1425C308.003 59.1425 308.799 59.282 309.41 59.4564C310.267 59.7354 316.445 60.8515 325.132 62.3164C326.172 62.4908 329.475 63.0838 332.472 63.6069C335.531 64.165 339.201 64.8277 340.731 65.0718C342.26 65.316 344.034 65.6299 344.707 65.7694C345.38 65.9089 348.683 66.5367 352.048 67.1296C355.412 67.7575 360.122 68.7689 362.447 69.3967C364.833 70.0246 367.83 70.8267 369.176 71.1755C370.522 71.5243 374.681 72.6055 378.352 73.617C382.083 74.6285 386.182 75.7097 387.528 76.0585C396.398 78.2558 409.917 82.1273 436.099 89.8703C442.338 91.7189 445.03 93.1489 447.721 96.1135C448.822 97.3692 450.902 99.6711 452.309 101.206C456.53 105.775 459.772 107.623 463.871 107.623C464.788 107.623 467.174 107.24 469.193 106.751C471.946 106.089 473.903 105.879 477.329 105.879C482.284 105.879 483.446 106.124 488.768 108.321C490.603 109.053 494.335 110.553 497.026 111.634C499.718 112.716 504 114.425 506.508 115.471C509.016 116.517 512.992 118.122 515.317 119.029C517.641 119.935 520.394 121.052 521.434 121.47C526.144 123.423 527.735 124.086 529.937 124.888C533.424 126.248 535.626 128.271 538.012 132.387C539.174 134.48 540.153 136.433 540.153 136.712C540.153 137.026 540.398 137.549 540.765 137.898C541.56 138.735 545.047 145.152 545.047 145.815C545.047 146.094 545.291 146.617 545.658 146.966C546.454 147.803 549.941 154.221 549.941 154.884C549.941 155.163 550.185 155.686 550.491 156.035C550.797 156.348 551.653 157.395 552.387 158.302C553.794 160.046 556.119 161.092 560.34 161.999C561.502 162.243 563.888 162.836 565.539 163.254C568.476 164.022 570.8 164.58 575.205 165.661C576.245 165.905 578.814 166.393 580.833 166.742C582.851 167.056 585.054 167.475 585.726 167.614C586.889 167.858 601.02 170.404 605.607 171.172C606.77 171.381 608.299 171.66 608.972 171.8C610.195 172.079 624.877 174.694 628.853 175.357C630.015 175.566 631.545 175.845 632.218 175.985C633.441 176.264 648.122 178.88 652.099 179.542C653.261 179.752 654.79 180.031 655.463 180.17C656.687 180.449 671.368 183.065 675.344 183.728C676.506 183.937 678.036 184.216 678.709 184.356C679.81 184.6 693.268 187.006 698.59 187.913C699.752 188.123 701.281 188.402 701.954 188.541C703.178 188.82 717.859 191.436 721.835 192.099C722.997 192.308 724.527 192.587 725.2 192.727C726.423 193.006 741.41 195.691 744.775 196.249C745.815 196.424 748.262 196.842 750.28 197.226C761.842 199.388 768.938 199.249 776.707 196.703C779.521 195.796 796.16 189.378 808.7 184.356C822.77 178.74 841.733 171.416 845.098 170.3C849.563 168.8 853.662 168.207 867.12 167.056C888.836 165.208 897.4 165.173 908.105 166.777C919.055 168.416 919.483 168.416 924.377 167.579C932.819 166.114 940.588 167.23 942.79 170.265C943.952 171.869 945.115 177.554 945.115 181.844C945.115 184.635 946.399 189.797 947.378 190.738C948.54 191.994 952.088 193.459 969.89 200.156C980.656 204.236 990.076 207.759 990.811 208.003C991.85 208.317 997.784 208.422 1014.06 208.422C1032.16 208.422 1036.38 208.526 1038.4 208.98C1039.75 209.259 1041.28 209.573 1041.77 209.677C1042.99 209.922 1045.44 212.119 1045.44 212.956C1045.44 213.305 1045.74 213.723 1046.05 213.828C1046.36 213.933 1046.66 214.665 1046.66 215.397C1046.66 216.165 1046.91 217.455 1047.27 218.257C1047.58 219.095 1048.31 220.978 1048.86 222.478C1049.6 224.64 1050.21 225.408 1051.92 226.419C1053.08 227.117 1054 227.988 1054 228.337C1054 229 1040.36 229 527.001 229H0L0 119.831Z"
-                      fill="#005837"
-                    />
-                    <g filter="url(#filter1_d_18_5179)">
-                      <path
-                        d="M1050.21 227.57C1048.37 226.419 1047.76 225.687 1046.97 223.489C1046.48 222.024 1045.74 220.141 1045.44 219.304C1045.07 218.502 1044.83 217.211 1044.83 216.444C1044.83 215.711 1044.58 214.979 1044.21 214.874C1043.91 214.77 1043.6 214.351 1043.6 214.002C1043.6 213.165 1041.16 210.968 1039.93 210.724C1039.44 210.619 1037.91 210.305 1036.57 210.026C1034.55 209.573 1030.33 209.468 1012.22 209.468C995.949 209.468 990.015 209.364 988.975 209.05C988.241 208.805 978.821 205.283 968.054 201.202C950.253 194.505 946.705 193.04 945.543 191.785C944.564 190.843 943.28 185.681 943.28 182.891C943.28 178.601 942.117 172.916 940.955 171.311C939.67 169.602 936.122 168.277 932.33 168.137C930.556 168.068 927.803 168.347 924.928 168.87C921.93 169.428 919.728 169.637 918.505 169.498C915.018 169.079 912.387 168.73 906.576 167.858C895.993 166.254 890.182 166.324 868.955 168.102C855.497 169.253 851.398 169.846 846.933 171.346C843.568 172.462 824.605 179.787 810.535 185.402C796.16 191.157 781.172 196.877 778.114 197.854C769.978 200.4 760.986 200.574 749.363 198.447C746.855 197.993 743.98 197.47 742.94 197.296C738.78 196.598 724.527 194.017 723.365 193.773C722.692 193.633 721.162 193.354 720 193.145C716.024 192.482 701.342 189.867 700.119 189.588C699.446 189.448 697.917 189.169 696.754 188.96C691.432 188.053 677.975 185.646 676.873 185.402C676.201 185.263 674.671 184.984 673.509 184.774C669.533 184.112 654.851 181.496 653.628 181.217C652.955 181.077 651.426 180.798 650.263 180.589C646.287 179.926 631.606 177.31 630.382 177.031C629.709 176.892 628.18 176.613 627.018 176.403C623.042 175.741 608.36 173.125 607.137 172.846C606.464 172.706 604.935 172.427 603.772 172.218C599.184 171.451 585.054 168.905 583.891 168.661C583.218 168.521 581.016 168.102 578.997 167.789C576.979 167.44 574.532 166.986 573.492 166.742C570.189 165.94 566.824 165.103 563.704 164.301C562.053 163.882 559.667 163.289 558.505 163.045C554.284 162.138 551.959 161.092 550.552 159.348C549.818 158.441 548.962 157.395 548.656 157.081C548.35 156.732 548.105 156.209 548.105 155.93C548.105 155.267 544.619 148.85 543.823 148.013C543.456 147.664 543.212 147.141 543.212 146.862C543.212 146.199 539.725 139.781 538.93 138.944C538.562 138.595 538.318 138.072 538.318 137.758C538.318 137.479 537.339 135.526 536.177 133.433C533.791 129.318 531.589 127.295 528.102 125.935C525.9 125.132 524.309 124.47 519.599 122.516C518.559 122.098 515.806 120.982 513.482 120.075C511.157 119.168 507.181 117.564 504.673 116.517C502.165 115.471 497.883 113.762 495.191 112.681C492.5 111.6 488.768 110.1 486.933 109.367C480.02 106.507 477.512 106.298 470.233 107.972C465.033 109.158 464.727 109.158 460.812 108.6C457.142 108.077 454.144 106.228 450.596 102.392C449.128 100.752 446.987 98.4156 445.886 97.1599C443.195 94.1953 440.503 92.7653 434.263 90.9167C408.082 83.1737 394.562 79.3022 385.692 77.1049C384.347 76.7561 380.248 75.6749 376.517 74.6634C372.846 73.6519 368.686 72.5707 367.341 72.2219C365.995 71.8731 362.997 71.071 360.612 70.4431C358.287 69.8153 353.577 68.8039 350.212 68.1761C346.848 67.5831 343.545 66.9553 342.872 66.8158C342.199 66.6763 340.425 66.3624 338.896 66.1182C337.366 65.8741 333.696 65.2114 330.637 64.6534C327.64 64.1302 324.336 63.5372 323.297 63.3628C314.61 61.8979 308.432 60.7818 307.575 60.5028C306.963 60.3284 306.168 60.1889 305.679 60.1889C305.251 60.1889 301.519 59.5611 297.298 58.7938C293.077 58.0265 288.856 57.2591 287.816 57.0847C286.838 56.9103 283.106 56.2825 279.558 55.6547C276.01 55.0618 272.646 54.3991 272.095 54.2247C271.483 54.0503 270.688 53.9108 270.199 53.9108C269.771 53.9108 266.039 53.283 261.818 52.5157C257.597 51.7484 253.376 50.981 252.336 50.8066C241.815 48.993 237.472 48.2256 236.432 47.9815C235.759 47.842 233.189 47.3886 230.743 46.9351C225.237 45.9934 222.668 44.668 219.609 41.1802C217.835 39.1224 215.449 36.6809 212.574 33.9604C207.375 29.0426 197.587 22.5552 192.387 20.602C189.39 19.4859 184.496 18.0907 180.765 17.3583C177.094 16.6259 170.121 16.2771 156.602 16.2073C149.016 16.1725 141.614 16.0329 140.085 15.8586C138.556 15.719 130.603 15.5795 122.345 15.5098C114.087 15.4749 106.134 15.3354 104.605 15.161C103.076 15.0215 95.1232 14.882 86.8649 14.8122C78.6066 14.7773 70.5319 14.6378 68.819 14.4634C67.1674 14.3239 58.4809 14.1146 49.5497 14.01C29.8522 13.7658 20.6151 13.3822 9.90994 12.4405C5.38318 12.0219 1.34579 11.708 0.856415 11.708C0.183517 11.708 0 10.6268 0 5.60432C0 1.62819 0.24469 -0.49939 0.611725 -0.49939C0.97876 -0.49939 3.548 -0.185476 6.30077 0.163307C22.4503 2.36064 25.4477 2.5699 45.9405 2.81405C54.9329 2.91869 73.774 3.16284 87.8437 3.33723C101.975 3.51162 126.015 3.82552 141.308 4.03479C176.605 4.48821 180.52 4.73235 190.308 7.20871C200.401 9.71995 207.252 12.7893 213.981 17.8117C221.322 23.2876 228.969 28.4845 231.354 29.6704C235.331 31.6236 238.083 32.1467 251.602 33.507C274.481 35.809 286.532 37.2738 290.263 38.2156C290.936 38.3899 292.894 38.7736 294.545 39.0875C296.258 39.4014 298.461 39.8548 299.439 40.1339C300.479 40.378 302.376 40.8315 303.721 41.1105C305.067 41.3895 308.799 42.3312 311.98 43.2032C321.155 45.6447 333.329 48.8883 337.366 49.8998C348.438 52.7598 355.29 52.5157 364.588 48.993C370.583 46.7259 372.418 46.5166 389.057 46.2376C426.372 45.54 440.992 45.4703 444.234 45.8888C450.352 46.7259 450.413 46.7607 472.863 57.9218C479.042 60.9911 485.22 64.0604 486.566 64.7231C487.973 65.3858 492.255 67.5134 496.109 69.4317C502.654 72.7102 509.077 75.5354 509.934 75.5354C510.301 75.5354 518.07 77.5583 528.836 80.4532C532.201 81.3252 536.054 82.3715 537.4 82.6854C538.746 83.0342 542.049 83.8713 544.741 84.6037C547.432 85.3362 552.143 86.5569 555.14 87.3243C560.952 88.824 567.681 90.5679 574.715 92.4514C577.101 93.0792 581.75 94.2999 585.115 95.1719C592.394 97.0204 594.107 97.4738 605.179 100.473C609.767 101.729 613.743 102.74 613.866 102.74C614.049 102.74 619.188 104.066 625.366 105.67C631.483 107.275 638.151 109.019 640.17 109.542C642.189 110.065 646.654 111.286 650.141 112.192C653.567 113.134 656.503 113.902 656.687 113.902C656.87 113.902 659.256 114.494 662.008 115.262C667.208 116.657 671.062 117.424 679.014 118.75C681.523 119.203 684.275 119.656 685.132 119.831C685.988 119.97 688.619 120.424 690.943 120.807C694.98 121.47 699.997 122.342 713.76 124.679C717.003 125.237 728.075 127.12 738.352 128.864C756.153 131.899 765.635 133.503 772.302 134.689C792 138.177 806.376 138.77 835.616 137.34C857.944 136.224 861.798 136.154 877.519 136.572C894.28 136.991 911.776 137.444 933.798 137.968C942.24 138.142 951.844 138.491 955.208 138.665C967.932 139.432 979.922 141.211 992.095 144.211C997.111 145.432 1018.15 151.605 1032.59 156.104C1035.47 156.976 1041.28 158.72 1045.44 159.906C1049.66 161.127 1053.39 162.243 1053.7 162.382C1054.06 162.557 1058.71 163.952 1064.1 165.521C1069.48 167.126 1074.13 168.556 1074.49 168.695C1074.8 168.87 1075.41 169.009 1075.84 169.009C1076.27 169.009 1078.41 169.498 1080.61 170.056C1082.81 170.649 1084.77 171.102 1084.89 171.102C1085.02 171.102 1086.97 171.59 1089.18 172.148C1091.38 172.741 1093.34 173.195 1093.46 173.195C1093.58 173.195 1095.54 173.683 1097.74 174.241C1099.94 174.834 1101.9 175.287 1102.02 175.287C1102.14 175.287 1104.04 175.741 1106.12 176.299C1108.26 176.822 1110.65 177.415 1111.5 177.554C1112.36 177.694 1114.01 178.113 1115.17 178.461C1116.34 178.845 1117.62 179.124 1117.99 179.124C1118.36 179.124 1120.31 179.577 1122.27 180.17C1124.23 180.728 1126.19 181.217 1126.61 181.217C1127.04 181.217 1128.57 181.6 1130.04 182.054C1131.57 182.507 1133.53 182.961 1134.44 183.1C1138.54 183.693 1143.86 186.588 1150.16 191.61C1156.89 196.982 1159.71 198.691 1163.99 199.981C1171.45 202.248 1179.96 201.795 1188.28 198.656C1195.8 195.831 1201.61 195.273 1209.07 196.738C1211.77 197.226 1215.74 198.133 1217.94 198.691C1220.15 199.284 1222.47 199.877 1223.14 200.051C1223.82 200.225 1226.2 200.818 1228.53 201.411C1230.79 202.004 1232.75 202.492 1232.93 202.492C1233.11 202.492 1235.38 203.051 1237.95 203.748C1240.58 204.446 1243.7 205.213 1244.86 205.457C1246.02 205.701 1247.86 206.155 1248.84 206.469C1249.88 206.783 1250.92 207.027 1251.22 207.027C1251.53 207.027 1253.49 207.48 1255.5 208.038C1257.58 208.596 1260.21 209.294 1261.38 209.573C1262.54 209.852 1265.11 210.515 1267 211.003C1268.9 211.491 1270.74 211.91 1271.04 211.91C1271.35 211.91 1272.82 212.258 1274.28 212.712C1275.75 213.165 1278.81 213.933 1281.07 214.456C1283.34 214.979 1286.27 215.711 1287.5 216.095C1288.78 216.479 1290.07 216.793 1290.37 216.793C1290.68 216.793 1292.64 217.281 1294.72 217.839C1305.05 220.699 1314.17 221.989 1329.58 222.722C1333.99 222.931 1341.51 223.35 1346.41 223.594C1355.58 224.117 1388.06 224.082 1423.97 223.559L1440 223.315V226.14V229H1246.27H1052.47L1050.21 227.57Z"
-                        fill="#00A850"
-                      />
-                    </g>
-                  </g>
-                  <defs>
-                    <filter
-                      id="filter0_d_18_5179"
-                      x="-1.94922"
-                      y="-3.9292"
-                      width="1443.95"
-                      height="306.941"
-                      filterUnits="userSpaceOnUse"
-                      color-interpolation-filters="sRGB"
-                    >
-                      <feFlood flood-opacity="0" result="BackgroundImageFix" />
-                      <feColorMatrix
-                        in="SourceAlpha"
-                        type="matrix"
-                        values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-                        result="hardAlpha"
-                      />
-                      <feOffset dy="-2" />
-                      <feGaussianBlur stdDeviation="1" />
-                      <feComposite in2="hardAlpha" operator="out" />
-                      <feColorMatrix
-                        type="matrix"
-                        values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-                      />
-                      <feBlend
-                        mode="normal"
-                        in2="BackgroundImageFix"
-                        result="effect1_dropShadow_18_5179"
-                      />
-                      <feBlend
-                        mode="normal"
-                        in="SourceGraphic"
-                        in2="effect1_dropShadow_18_5179"
-                        result="shape"
-                      />
-                    </filter>
-                    <filter
-                      id="filter1_d_18_5179"
-                      x="-4"
-                      y="-0.49939"
-                      width="1448"
-                      height="237.499"
-                      filterUnits="userSpaceOnUse"
-                      color-interpolation-filters="sRGB"
-                    >
-                      <feFlood flood-opacity="0" result="BackgroundImageFix" />
-                      <feColorMatrix
-                        in="SourceAlpha"
-                        type="matrix"
-                        values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-                        result="hardAlpha"
-                      />
-                      <feOffset dy="4" />
-                      <feGaussianBlur stdDeviation="2" />
-                      <feComposite in2="hardAlpha" operator="out" />
-                      <feColorMatrix
-                        type="matrix"
-                        values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"
-                      />
-                      <feBlend
-                        mode="normal"
-                        in2="BackgroundImageFix"
-                        result="effect1_dropShadow_18_5179"
-                      />
-                      <feBlend
-                        mode="normal"
-                        in="SourceGraphic"
-                        in2="effect1_dropShadow_18_5179"
-                        result="shape"
-                      />
-                    </filter>
-                    <clipPath id="clip0_18_5179">
-                      <rect width="1440" height="229" fill="white" />
-                    </clipPath>
-                  </defs>
-                </svg>
-              </div>
-            </div>
-          </div>
-          <div className="page-buttons">
-            <div className="image-home">
-              <img
-                src="https://images.unsplash.com/photo-1597655601841-214a4cfe8b2c?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Mnx8bW91bnRhaW4lMjBzY2VuZXJ5fGVufDB8fDB8fHww&w=1000&q=80"
-                style={{ width: "100%", height: "100%" }}
-              ></img>
-            </div>
-            <div className="text-and-buttons">
-              <p className="image-description">
-                Nestled in the heart of Ontarios picturesque Niagara Region,
-                Lincoln is a charming and vibrant community that often flies
-                under the radar. Lincoln offers its own unique blend of natural
-                beauty, rich history, and thriving agriculture. <br />
-                <br />
-                On a journey through Lincoln, you will uncover attractions,
-                culture, and events that make it a hidden gem worth discovering
-              </p>
-              <div className="buttons">
-                <DefaultButton
-                  isLink={true}
-                  href="./events"
-                  className=" hero-under-buttons"
-                >
-                  Events
-                </DefaultButton>
-                <DefaultButton
-                  isLink={true}
-                  href="./attractions"
-                  className=" hero-under-buttons"
-                >
-                  Attractions
-                </DefaultButton>
-                <DefaultButton
-                  isLink={true}
-                  href="./business-service"
-                  className=" hero-under-buttons"
-                >
-                  Businesses
-                </DefaultButton>
-              </div>
-            </div>
-          </div>
-
-          <Section marginBottom="40px" marginTop="200px">
-            <CardCarousel title="Events" margin="80px 0px 80px 0px">
-              {eventsStrapiData?.map((card, index) => (
-                <li key={index}>
-                  <LargeCardMobile
-                    isTicket={false}
-                    isEvent={true}
-                    title={card.attributes.title}
-                    description={
-                      card.attributes.richTextDescription != undefined
-                        ? card.attributes.richTextDescription
-                        : card.attributes.description
-                    }
-                    address={card.attributes.location}
-                    ticketDate={card.attributes.date}
-                    ticketTime={`${card.attributes.startTime} - ${card.attributes.endTime}`}
-                    rating={card.attributes.numStars}
-                    category={card.attributes.tags}
-                    imgSrc={card.attributes.image.data.attributes.url}
-                    imgAltText={
-                      card.attributes.image.data.attributes.alternativeText
-                    }
-                    barcodeUID={card.attributes.barcodeUID}
-                    isRegisterable={card.attributes.isRegisterable}
-                    isFull={card.attributes.isFull}
-                    isAvail={card.attributes.isAvailable}
-                    hoursOfOperation={card.attributes.hoursOfOperation}
-                  ></LargeCardMobile>
-                </li>
-              ))}
-            </CardCarousel>
-            <DefaultButton
-              isLink={true}
-              href="./events"
-              className=" homepage-see-more"
-            >
-              See More
-            </DefaultButton>
-            <div className="carousel-padding"></div>
-            <hr />
-            <div className="carousel-padding"></div>
-            <CardCarousel title="Attractions" margin="0px 0px 80px 0px">
-              {attractionStrapiData?.map((card, index) => (
-                <li key={index}>
-                  <LargeCardMobile
-                    isTicket={false}
-                    isEvent={false}
-                    title={card.attributes.title}
-                    description={
-                      card.attributes.richTextDescription != undefined
-                        ? card.attributes.richTextDescription
-                        : card.attributes.description
-                    }
-                    address={card.attributes.location}
-                    ticketDate={card.attributes.date}
-                    ticketTime={`${card.attributes.startTime} - ${card.attributes.endTime}`}
-                    rating={card.attributes.numStars}
-                    category={card.attributes.tags}
-                    imgSrc={card.attributes.image.data.attributes.url}
-                    imgAltText={
-                      card.attributes.image.data.attributes.alternativeText
-                    }
-                    barcodeUID={card.attributes.barcodeUID}
-                    isRegisterable={card.attributes.isRegisterable}
-                    isFull={card.attributes.isFull}
-                    isAvail={card.attributes.isAvailable}
-                    hoursOfOperation={card.attributes.hoursOfOperation}
-                  ></LargeCardMobile>
-                </li>
-              ))}
-            </CardCarousel>
-            <DefaultButton
-              isLink={true}
-              href="./attractions"
-              className=" homepage-see-more"
-            >
-              See More
-            </DefaultButton>
-            <div className="carousel-padding"></div>
-            <hr />
-            <div className="carousel-padding"></div>
-
-            <CardCarousel title="Business" margin="20px 0px 80px 0px">
-              {businessStrapiData?.map((card, index) => (
-                <li key={index}>
-                  <LargeCardMobile
-                    isTicket={false}
-                    isEvent={false}
-                    title={card.attributes.title}
-                    description={
-                      card.attributes.richTextDescription != undefined
-                        ? card.attributes.richTextDescription
-                        : card.attributes.description
-                    }
-                    address={card.attributes.location}
-                    ticketDate={card.attributes.date}
-                    ticketTime={`${card.attributes.startTime} - ${card.attributes.endTime}`}
-                    rating={card.attributes.numStars}
-                    category={card.attributes.tags}
-                    imgSrc={card.attributes.image.data.attributes.url}
-                    imgAltText={
-                      card.attributes.image.data.attributes.alternativeText
-                    }
-                    barcodeUID={card.attributes.barcodeUID}
-                    isRegisterable={card.attributes.isRegisterable}
-                    isFull={card.attributes.isFull}
-                    isAvail={card.attributes.isAvailable}
-                    hoursOfOperation={card.attributes.hoursOfOperation}
-                  ></LargeCardMobile>
-                </li>
-              ))}
-            </CardCarousel>
-            <DefaultButton
-              isLink={true}
-              href="./business-service"
-              className=" homepage-see-more"
-            >
-              See More
-            </DefaultButton>
-
-            <div className="carousel-padding"></div>
-            <hr />
-            <div className="carousel-padding"></div>
-
-            <h2 className="events-title">Gallery</h2>
-
-            <Gallery>
-              {galleryStrapiData?.map((image, index) => (
-                <GalleryImage
-                  src={image.attributes.url}
-                  alt={
-                    "https://travelforfoodhub.com/wp-content/uploads/2023/05/Best-Wine-Regions-in-Europe.jpg"
-                  }
-                  key={index}
-                />
-              ))}
-            </Gallery>
-          </Section>
-        </div>
-
-        <Footer></Footer>
-      </div>
-
-      <style jsx>{`
-        li {
-          margin: 20px 20px;
-        }
-
-        .smallCardli {
-          margin: 10px 00px 5px 0px;
-        }
-
-        h2 {
-          color: var(--color-font-primary);
-          font-family: var(--font-calps);
-          font-size: var(--font-size-header-S);
-          font-style: normal;
-          font-weight: 700;
-          line-height: normal;
-        }
-
-        .homepage {
-          display: flex;
-          flex-direction: column;
-
-          height: 100vh;
-          justify-content: space-between;
-        }
-
-        .landing-view-bg {
-          display: flex;
-          flex-direction: column;
-          background-image: url('https://cdn.discordapp.com/attachments/986414801448235038/1161769104684896366/kelsey-knight-udj2tD3WKsY-unsplash_2.png?ex=653980cb&is=65270bcb&hm=7aa2d5ba68cfdfbf6dbaa77b361365cda1e859a0ce094d587331f5a8f253e4bb&');
-          background-size: cover;
-          background-repeat: no-repeat;
-        }
-
-        .landing-view-greet {
-          display: flex;
-          flex-direction: row;
-          height: auto;
-          justify-content: space-evenly;
-          align-items: center;
-          width: 100%;
-          padding-top: 75px;
-        }
-
-        .greeting-frame {
-          display: flex;
-          flex-flow: column wrap;
-          align-content: flex-start;
-          margin-left: auto;
-          margin-right: auto;
-        }
-
-        .discover-text {
-          font-size: 73px;
-          font-family: var(--font-calps);
-          font-style: normal;
-          font-weight: 700;
-          line-height: 75.1%;
-          color: white;
-          padding-bottom: 10px;
-          text-size-adjust: auto;
-        }
-
-        .lincoln-text {
-          font-size: 125px;
-          font-family: var(--font-calps);
-          font-style: normal;
-          font-weight: 950;
-          line-height: 75.1%; /* 93.875px */
-          text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-
-          color: white;
-        }
-        .greeting-description-text {
-          font-family: var(--font-roboto);
-          font-style: normal;
-          font-weight: 500;
-          font-size: 24;
-          color: white;
-          max-width: 550px;
-          text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-          line-height: 150%;
-          padding-top: 20px;
-          padding-bottom: 35px;
-        }
-
-        .search-area {
-          display: flex;
-          flex-direction: row;
-          justify-content: flex-start;
-          padding: 0px 0px 0px 0px;
-          max-height: 40px;
-        }
-
-        .search-button-area {
-          display: flex;
-          flex-direction: row;
-          justify-content: flex-start;
-          padding: 0px 0px 0px 20px;
-        }
-
-        .upcoming-events-gap-desktop {
-          display: flex;
-          height: 40px;
-        }
-
-        .upcoming-events-gap-mobile {
-          display: none;
-        }
-
-        .upcoming-events-wrap {
-          display: flex;
-          flex-flow: column wrap;
-          align-content: flex-start;
-          max-width: 320px;
-          height: auto;
-          margin-top: 40px;
-          margin-left: auto;
-          margin-right: auto;
-        }
-
-        .upcoming-events-subtitle {
-          color: #fff;
-          text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-          font-family: var(--font-roboto);
-          font-size: 24px;
-          font-style: normal;
-          font-weight: 700;
-          line-height: 75.1%; /* 18.024px */
-          margin-bottom: 20px;
-        }
-
-        .events-card-frame {
-          width: 393px;
-          height: 429px;
-          background: rgb(220, 20, 60, 0.6);
-          padding: 18px 15px 0px 16px;
-          flex-direction: column;
-          justify-content: flex-end;
-          align-items: flex-start;
-          box-shadow: 3px 5px 4px 0px rgba(0, 0, 0, 0.35);
-        }
-
-        .see-all-events-button {
-          display: flex;
-          width: 83px;
-          height: 34px;
-          padding: 7px 2px 2px 12px;
-          background: #fff;
-          backdrop-filter: blur(2px);
-          color: black;
-          font-family: var(--font-roboto);
-          font-size: 16px;
-          font-weight: 800;
-          align-self: flex-end;
-        }
-        .page-buttons {
-          display: flex;
-          flex-direction: row;
-          flex-wrap: wrap;
-
-          background: var(--color-elevated-green);
-          justify-content: center;
-          width: 100%;
-          gap: 30px;
-          align-items: center;
-        }
-
-        .topology-wrap {
-          display: flex;
-          flex-direction: column;
-        }
-
-        .image-home {
-          height: 430px;
-          width: 390px;
-          border: 5px solid white;
-          position: relative;
-          margin-bottom: -100px;
-          box-shadow: var(--shadow-box-buttons);
-        }
-
-        .image-home img {
-          object-fit: cover;
-        }
-
-        .text-and-buttons {
-          display: flex;
-          flex-direction: column;
-          max-width: 900px;
-          padding-left: 20px;
-        }
-
-        .carousel-padding {
-          height: 30px;
-        }
-
-        .events-title{
-          padding: 20px 0px;
-        }
-
-        .image-description {
-          font-family: var(--font-roboto);
-          font-style: normal;
-          font-size: 18px;
-          color: white;
-          text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-          line-height: 170%;
-          padding-bottom: 35px;
-          padding-top: 20px;
-        }
-
-        .buttons {
-          display: flex;
-          flex-direction: row;
-          margin-bottom: 30px;
-          gap: 15px;
-        }
-
-        .buttons button {
-          display: flex;
-          padding: 6px 15px;
-          background: var(--color-topographic-green);
-          backdrop-filter: blur(2px);
-          color: white;
-          font-family: var(--font-roboto);
-          font-size: 14px;
-          font-weight: 800;
-          font-style: normal;
-          font-weight: 600;
-          line-height: 140.1%; /* 22.416px */
-          margin-left: 10px;
-          margin-right: 10px;
-          text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-        }
-
-
-        .see-lincoln-wrapper {
-          display: flex;
-          flex-direction: row;
-          align-items: stretch;
-        }
-
-        @media screen and (max-width: 3000px) {
-          .image-description {
-            font-size: 24;
-            width: 350px;
-          }
-        }
-
-        @media screen and (max-width: 870px) {
-          .landing-view-greet {
-            margin-top: 120px;
-            height: 1100px;
-            flex-wrap: wrap;
-            gap: 20px;
-            height: auto;
-          }
-
-          .image-home {
-            height: 430px;
-            width: auto;
-            max-width: 40%;
-            margin-left: 20px;
-          }
-        }
-
-        @media screen and (max-width: 720px) {
-          .image-home {
-            height: 230px;
-            width: auto;
-            max-width: 100%;
-            margin-left: 20px;
-            margin-right: 20px;
-            margin-bottom: 0px;
-            margin-top: 20px;
-          }
-        }
-
-        @media screen and (max-width: 550px) {
-          .discover-text {
-			    	font-size: 64px;
-			    }
-
-          .lincoln-text {
-            font-size: 82px;
-          }
-
-          .see-lincoln-wrapper {
-            display: none;
-          }
-
-          .greeting-frame {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-          }
-        }
-      `}</style>
-    </>
-  );
-}
-
-Homepage.overrideDefaultNavs = true;
